src/util.py: search for iptables when exe is not specified
We prefer to hardcode the iptables dir in common.py, but we do not
import common.py here. While internally ufw always uses common.py to
determine the path, _find_system_iptables() is implemented for
get_iptables_version() and get_netfilter_capabilities() so as to not
break API for external consumers since these have historically used a
default for 'exe'.