detect if should use nft or legacy
iptables 1.8 supports using either the nft backend or the historic
legacy (xtables) backend. Software adding rules must consistently use
the same backend otherwise the kernel may apply the rules in an
unpredictable (to the admin) fashion.

ufw via distro-packaging or setup.py install will use iptables/ip6tables
from the system, but ufw as a snap needs to choose the most appropriate
backend to use at runtime since it may be installed anywhere.

Add detect-iptables to manage symlinks in SNAP_DATA, and use it in
bin/srv and in the install hook. The algorithm for choosing which
backend to use is:

* if kernel < 5.0, use legacy
* if only legacy in use, then use it. Otherwise, prefer nft for:
  - neither in use
  - both in use
  - nft in use

Update ufw via Makefile (for now) to use iptables commands via the
managed symlinks in SNAP_DATA.

Unfortunately, iptables -L detects if there are legacy rules by checking
if the iptable_filter (same with ip6tables and ip6table_filter), so the
act of checking if legacy is in use loads this module (and snaps aren't
allowed to unload the module) and iptables-nft -L will output:

 # Warning: iptables-legacy tables present, use iptables-legacy to see them

Short of unloading the module, there is nothing the snap is able to do
to get rid of this message. A future commit will adjust 'ufw show raw'
output to filter out this message.