debian/control: Breaks with iptables-persistent and netfilter-persistent
When ufw is installed, it is not enabled by default, so it doesn't
interfere with other firewall software (until it is enabled). In
contrast, iptables-persistent and netfilter-persistent install enabled,
which interferes with ufw. Add a breaks on these to avoid them being
co-installed with ufw (and causing problems for users).